requirements.txt
streamlit
torch
transformers
numpy
pillow
matplotlib
evaluate
tqdm